requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
aiosqlite==0.17.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asyncpg==0.30.0
asyncpraw==7.8.1
asyncprawcore==2.4.0
attrs==25.3.0
beautifulsoup4==4.13.4
blis==1.3.0
catalogue==2.0.10
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cssselect==1.3.0
cymem==2.0.11
distro==1.9.0
editdistpy==0.1.6
fastapi==0.116.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
frozenlist==1.7.0
greenlet==3.2.3
griffe==1.7.3
groq==0.30.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
iniconfig==2.1.0
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.0
lxml_html_clean==0.4.2
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.12.2
mdurl==0.1.2
multidict==6.6.3
murmurhash==1.0.13
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.3.2
openai==1.98.0
openai-agents==0.2.4
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
praw==7.8.1
prawcore==2.4.0
preshed==3.0.10
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.31
requests==2.32.4
requests-file==2.1.0
rich==14.0.0
rpds-py==0.26.0
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.42
srsly==2.5.1
sse-starlette==2.4.1
starlette==0.47.1
symspellpy==6.9.0
tenacity==9.1.2
thinc==8.3.6
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
update-checker==0.18.0
urllib3==2.5.0
uv==0.7.19
uvicorn==0.35.0
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
wrapt==1.17.2
yarl==1.20.1